function that searches for a value in an array of integers using the Linear search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,27 +1,27 @@
 #include "search_algos.h"
 
 /**
-  * linear_search - Searches for a val in an array
-  *                 of ints using linear search.
-  * @array: A pointer to the first element of the search array.
-  * @size: Array elements.
-  * @value: Search value.
+  * linear_search - Searches for a value in an array
+  *                 of integers using linear search.
+  * @array: A pointer to the first element of the array to search.
+  * @size: The number of elements in the array.
+  * @value: The value to search for.
   *
   * Return: If the value is not present or the array is NULL, -1.
-  *         else, the first index where the value is located.
+  *         Otherwise, the first index where the value is located.
   *
   * Description: Prints a value every time it is compared in the array.
   */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (array == NULL)
 		return (-1);
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%u] = [%d]\n", i, array[i]);
+		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
